pages/requirement.txt

(safa-mG-gM-Kg) (base) C:\safa>pipenv requirements
-i https://pypi.org/simple
affine==2.4.0 ; python_version >= '3.7'
altair==5.1.2 ; python_version >= '3.8'
aniso8601==9.0.1
anywidget==0.7.1 ; python_version >= '3.7'
asttokens==2.4.1
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
blinker==1.7.0 ; python_version >= '3.8'
bqplot==0.12.42 ; python_version >= '3.6'
branca==0.7.0 ; python_version >= '3.7'
cachelib==0.9.0 ; python_version >= '3.7'
cachetools==5.3.2 ; python_version >= '3.7'
certifi==2023.11.17 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
click-plugins==1.1.1
cligj==0.7.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3' and python_version < '4'
colorama==0.4.6 ; platform_system == 'Windows'
colour==0.1.5
comm==0.2.0 ; python_version >= '3.8'
contourpy==1.2.0 ; python_version >= '3.9'
cramjam==2.7.0 ; python_version >= '3.7'
cycler==0.12.1 ; python_version >= '3.8'
decorator==5.1.1 ; python_version >= '3.5'
deprecation==2.1.0
duckdb==0.9.2 ; python_full_version >= '3.7.0'
earthengine-api==0.1.379 ; python_version >= '3.7'
eerepr==0.0.4 ; python_version >= '3.7'
exceptiongroup==1.2.0 ; python_version < '3.11'
executing==2.0.1 ; python_version >= '3.5'
fastparquet==2023.10.1
filelock==3.13.1 ; python_version >= '3.8'
fiona==1.9.5 ; python_version >= '3.7'
flask==2.3.3 ; python_version >= '3.8'
flask-caching==2.1.0 ; python_version >= '3.7'
flask-cors==4.0.0
flask-restx==1.2.0
folium==0.14.0
fonttools==4.45.1 ; python_version >= '3.8'
fsspec==2023.10.0 ; python_version >= '3.8'
future==0.18.3 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gdown==4.7.1
geemap==0.29.6
geocoder==1.38.1
geojson==3.1.0 ; python_version >= '3.7'
geopandas==0.14.1
gitdb==4.0.11 ; python_version >= '3.7'
gitpython==3.1.40 ; python_version >= '3.7'
google-api-core==2.14.0 ; python_version >= '3.7'
google-api-python-client==2.108.0 ; python_version >= '3.7'
google-auth==2.23.4 ; python_version >= '3.7'
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3 ; python_version >= '3.7'
google-cloud-storage==2.13.0 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.6.0 ; python_version >= '3.7'
googleapis-common-protos==1.61.0 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
httplib2==0.22.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.6 ; python_version >= '3.5'
imageio==2.33.0
importlib-metadata==6.8.0 ; python_version >= '3.8'
importlib-resources==6.1.1 ; python_version >= '3.8'
ipyevents==2.0.2 ; python_version >= '3.6'
ipyfilechooser==0.6.0
ipyleaflet==0.17.4 ; python_version >= '3.7'
ipython==8.18.0 ; python_version >= '3.9'
ipytree==0.2.2
ipywidgets==8.1.1 ; python_version >= '3.7'
itsdangerous==2.1.2 ; python_version >= '3.7'
jedi==0.19.1 ; python_version >= '3.6'
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.17.3 ; python_version >= '3.7'
jupyterlab-widgets==3.0.9 ; python_version >= '3.7'
kiwisolver==1.4.5 ; python_version >= '3.7'
large-image[rasterio]==1.26.1 ; python_version >= '3.8'
large-image-source-rasterio==1.26.1
leafmap==0.29.3
git+https://github.com/banesullivan/localtileserver.git@41aed4585e2a833b7ce7f97f4613e34dff79dc44#egg=localtileserver
lonboard==0.4.2
markdown-it-py==3.0.0 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
matplotlib==3.8.2 ; python_version >= '3.9'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
mdurl==0.1.2 ; python_version >= '3.7'
mplleaflet==0.0.5
mypy-extensions==1.0.0 ; python_version >= '3.5'
numpy==1.26.2 ; python_version >= '3.9'
packaging==23.2 ; python_version >= '3.7'
palettable==3.3.3 ; python_version >= '3.7'
pandas==2.1.3 ; python_version >= '3.9'
parso==0.8.3 ; python_version >= '3.6'
pillow==10.1.0 ; python_version >= '3.8'
plotly==5.18.0 ; python_version >= '3.6'
prompt-toolkit==3.0.41 ; python_full_version >= '3.7.0'
protobuf==4.25.1 ; python_version >= '3.8'
psygnal==0.9.5 ; python_version >= '3.7'
pure-eval==0.2.2
pyarrow==14.0.1 ; python_version >= '3.8'
pyasn1==0.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydeck==0.8.0
pygments==2.17.2 ; python_version >= '3.7'
pyparsing==3.1.1 ; python_full_version >= '3.6.8'
pyperclip==1.8.2
pyproj==3.6.1 ; python_version >= '3.9'
pyrsistent==0.20.0 ; python_version >= '3.8'
pyshp==2.3.1 ; python_version >= '2.7'
pysocks==1.7.1
pystac[validation]==1.8.3 ; python_version >= '3.8'
pystac-client==0.7.5 ; python_version >= '3.8'
python-box==7.1.1 ; python_version >= '3.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
rasterio==1.3.9
ratelim==0.1.6
requests==2.31.0 ; python_version >= '3.7'
rich==13.7.0 ; python_full_version >= '3.7.0'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
scooby==0.9.2 ; python_version >= '3.8'
server-thread==0.2.0 ; python_version >= '3.7'
setuptools==69.0.2 ; python_version >= '3.8'
shapely==2.0.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1 ; python_version >= '3.7'
snuggs==1.4.7
soupsieve==2.5 ; python_version >= '3.8'
stack-data==0.6.3
streamlit==1.28.2
streamlit-folium==0.16.0
streamlit-modal==0.1.0
tenacity==8.2.3 ; python_version >= '3.7'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.3.3 ; python_version >= '3.8'
tqdm==4.66.1 ; python_version >= '3.7'
traitlets==5.13.0 ; python_version >= '3.8'
traittypes==0.2.1
typing-extensions==4.8.0 ; python_version >= '3.8'
tzdata==2023.3 ; python_version >= '2'
tzlocal==5.2 ; python_version >= '3.8'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==2.1.0 ; python_version >= '3.8'
uvicorn==0.24.0.post1 ; python_version >= '3.8'
validators==0.22.0 ; python_version >= '3.8'
vega-datasets==0.9.0
watchdog==3.0.0 ; platform_system != 'Darwin'
wcwidth==0.2.12
werkzeug==2.3.8 ; python_version >= '3.8'
whitebox==2.3.1
whiteboxgui==2.3.0 ; python_version >= '3.5'
widgetsnbextension==4.0.9 ; python_version >= '3.7'
xyzservices==2023.10.1 ; python_version >= '3.8'
zipp==3.17.0 ; python_version >= '3.8'
dem;